refactor(exception): exception 도메인 분리
