+++ src/main/java/nextstep/subway/common/exception/pathFinder/DisconnectedPathException.java
@@ -1,4 +1,6 @@
-package nextstep.subway.common.exception;
+package nextstep.subway.common.exception.pathFinder;
+
+import nextstep.subway.common.exception.BusinessException;
 
 public class DisconnectedPathException extends BusinessException {
     public DisconnectedPathException() {
