+++ src/main/java/nextstep/subway/common/exception/station/NotExistStationException.java
@@ -1,0 +1,9 @@
+package nextstep.subway.common.exception.station;
+
+import nextstep.subway.common.exception.BusinessException;
+
+public class NotExistStationException extends BusinessException {
+    public NotExistStationException() {
+        super("역이 존재하지 않습니다.");
+    }
+}
