+++ src/main/java/nextstep/subway/domain/Sections.java
@@ -1,8 +1,8 @@
 package nextstep.subway.domain;
 
-import nextstep.subway.common.exception.ContainsAllStationException;
-import nextstep.subway.common.exception.NotExistStationException;
-import nextstep.subway.common.exception.OnlyOneSectionException;
+import nextstep.subway.common.exception.line.ContainsAllStationException;
+import nextstep.subway.common.exception.station.NotExistStationException;
+import nextstep.subway.common.exception.line.OnlyOneSectionException;
 
 import javax.persistence.*;
 import java.util.*;
